+++ AMAPVox/AMAPVoxCore/src/main/java/fr/amap/lidar/amapvox/simulation/transmittance/lai2xxx/Lai2xxxSim.java
@@ -259,7 +259,7 @@ public class Lai2xxxSim implements Cancellable{
                             break;
                         }
 
-                        if(Float.isNaN(voxel.PadBVTotal)){
+                        if(Double.isNaN(voxel.PadBVTotal)){
                             gotOneNaN = true;
                             break;
                         }
